Added more art & alterations
Added art for the upper and lower frames for the main game scene as well as borders for the score and the target score texts.

Altered the graphics for the gumdrop images so they look more visible when entering the screen.

Changed the position of the timer as well as made adjustments to the score texts.

Changed the way the object pooler spawns in objects on the y axis based on the current active scene.