Script edits, bug fixes, & art changes
Edited the target color script such that when the gumdrop color is about to change it will select a color from any of the current gumdrops active in the scene at random.

Fixed a bug where the gumdrop speed would not reset to its correct incremental speed when the slow power up ends as the next stage begins.

Changed the gradient of the background slightly.