Bug Fixes, animations, & Changes
Fixed a bug where if a gumdrop of the target gumdrop color is hit right before the next color changes, that gumdrop would freeze and stay suspended in the game.

Fixed a bug where when opening the menu would show the menu's texts before opening when touched the first time.

Added a color animation for when the player gains or losses a score.

Gumdrops that are hit of the wrong color play an error sound effect.