Bonus score, adjustments & Sound effects
Changed the high score frame into the score goal frame that will glow once the player reaches that point.

Added a bonus score that the player can receive if they hit a certain number of the correct gumdrops in a row.

Added sound effects for the bonus score and for reaching the score goal.